Split stdlib template module
